+++ netharn/metrics/sklearn_alts.py
@@ -27,9 +27,9 @@ def confusion_matrix(y_true, y_pred, n_labels=None, labels=None,
         >>> y_pred = np.array([0, 0, 0, 0, 0, 0, 0, 1,  1, 1])
         >>> confusion_matrix(y_true, y_pred, 2)
         array([[4, 2],
-               [3, 1]])
+               [3, 1]]...)
         >>> confusion_matrix(y_true, y_pred, 2).ravel()
-        array([4, 2, 3, 1])
+        array([4, 2, 3, 1]...)
 
     Benchmarks:
         import ubelt as ub