+++ CHANGELOG.md
@@ -4,7 +4,17 @@ This changelog follows the specifications detailed in: [Keep a Changelog](https:
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html), although we have not yet reached a `1.0.0` release.
 
 
-## Version 0.5.15 - Unreleased
+## Version 0.5.16 - Unreleased
+
+### Added
+* decollate batch utility
+
+### Fixed
+* Fixed segmentation, and object detection example
+* Fixed issue with container_abcs
+
+
+## Version 0.5.15 - Released 2021-01-26
 
 
 ## Version 0.5.14 - Released 2021-01-21